♻️ Refactor :  shadow 스타일 코드 개선
