+++ src/feature/brand/ui/organisms/SelectedBrand.tsx
@@ -4,7 +4,7 @@ import { View, Text, Pressable } from 'react-native';
 import { ScrollView } from 'react-native-gesture-handler';
 
 import CloseIcons from '@/shared/icons/CloseIcons';
-import { boxShadow } from '@/styles/boxShadow';
+import { insetShadow } from '@/styles/shadows';
 
 interface Props {
   selectedList: {
@@ -36,7 +36,7 @@ const SelectedBrand = ({ selectedList, onPressIn }: Props) => {
             <View
               key={item.brandId}
               className="ml-5 flex-row items-center justify-center rounded-[27px] bg-pink-500 pb-2 pl-3 pr-2 pt-2"
-              style={boxShadow}>
+              style={{ boxShadow: insetShadow.default }}>
               <Text className="mr-1 text-white headline-01">{item.name}</Text>
 
               <Pressable onPressIn={() => onPressIn(item.brandId)}>
