+++ src/feature/map/ui/organisms/CurrentLocationSearch.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Pressable, Text, View } from 'react-native';
 
 import ReplayIcons from '@/shared/icons/ReplayIcon';
-import { commonShadow } from '@/styles/shadow';
+import { defaultShadow } from '@/styles/shadows';
 
 interface Props {
   showSearchButton: boolean;
@@ -22,7 +22,7 @@ const CurrentLocationSearch = ({
         <Pressable
           className="h-[40px] w-[149px] flex-shrink-0 flex-row items-center justify-center self-center rounded-[27px] bg-neutral-white"
           onPress={onLocationSearch}
-          style={commonShadow}>
+          style={defaultShadow}>
           <ReplayIcons height={24} width={24} shape="default" />
           <Text className="ml-1 text-semantic-info body-rg-02">
             현 지도에서 검색
