+++ src/shared/ui/molecules/SocialLoginButton/index.tsx
@@ -6,7 +6,7 @@ import { SocialTypes } from '@/feature/auth/login/types';
 import LoginTooltip from '@/feature/auth/login/ui/organisms/LoginTooltip';
 import SocialButtons from '@/shared/icons/SocialButtons';
 import { useUserStore } from '@/shared/store';
-import { commonShadow } from '@/styles/shadow';
+import { defaultShadow } from '@/styles/shadows';
 
 interface Props {
   type: SocialTypes;
@@ -21,7 +21,7 @@ const SocialLoginButton = ({ type, onPressIn }: Props) => {
       <Pressable
         onPressIn={() => onPressIn(type)}
         className="items-center justify-center"
-        style={commonShadow}>
+        style={defaultShadow}>
         <SocialButtons shape={type} />
       </Pressable>
