--- src/styles/bottomSheetShadow.ts
@@ -1,7 +1,0 @@
-export const bottomSheetShadow = {
-  shadowColor: '#000',
-  shadowOffset: { width: 0, height: -2 },
-  shadowOpacity: 0.2,
-  shadowRadius: 8,
-  elevation: 6,
-};
