--- src/styles/boxShadow.ts
@@ -1,4 +1,0 @@
-export const boxShadow = {
-  boxShadow:
-    '-2px -4px 12px 0px rgba(255, 255, 255, 0.10) inset, -2px -2px 8px 0px rgba(0, 0, 0, 0.20) inset, 2px 4px 8px 0px rgba(255, 255, 255, 0.25) inset',
-};
