--- src/styles/filterButtonShadow.ts
@@ -1,12 +1,0 @@
-export const filterButtonShadow = {
-  inactive: {
-    boxShadow: '0px 2px 6px 0px rgba(0, 0, 0, 0.20)',
-  },
-  active: {
-    boxShadow:
-      '0px 2px 4px 0px rgba(0, 0, 0, 0.20), ' +
-      '-2px -4px 12px 0px rgba(255, 255, 255, 0.10) inset, ' +
-      '-2px -2px 8px 0px rgba(0, 0, 0, 0.20) inset, ' +
-      '2px 4px 8px 0px rgba(255, 255, 255, 0.25) inset',
-  },
-};
