--- src/styles/inputShadow.ts
@@ -1,7 +1,0 @@
-export const inputShadow = {
-  shadowColor: '#000',
-  shadowOffset: { width: 0, height: 2 },
-  shadowOpacity: 0.2,
-  shadowRadius: 3,
-  elevation: 4,
-};
