--- src/styles/shadow.ts
@@ -1,7 +1,0 @@
-export const commonShadow = {
-  shadowColor: '#000',
-  shadowOffset: { width: 2, height: 2 },
-  shadowOpacity: 0.2,
-  shadowRadius: 2,
-  elevation: 6,
-};
